# src/main/java/org/example/dto/CartItemRequestDto.java
package org.example.dto;

import jakarta.validation.constraints.Max;
import jakarta.validation.constraints.Min;
import jakarta.validation.constraints.NotNull;
import lombok.Data;

@Data
public class CartItemRequestDto {
    @NotNull
    private Long bookId;
    @Min(1)
    @Max(100)
    private int quantity;
}