# src/main/resources/openAPI/specification/shopping_cart-manager-v1.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Shopping Cart API",
    "version": "1.0.0",
    "description": "API for managing shopping carts"
  },
  "servers": [
    {
      "url": "https://example.com/api"
    }
  ],
  "paths": {
    "/cart": {
      "post": {
        "summary": "Add a book to the shopping cart",
        "tags": ["Shopping Cart"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CartItemRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Book added to the cart successfully"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "get": {
        "summary": "Get all items in the shopping cart",
        "tags": ["Shopping Cart"],
        "responses": {
          "200": {
            "description": "List of items in the cart",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShoppingCartDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/cart/cart-items/{cartItemId}": {
      "put": {
        "summary": "Update the quantity of a cart item",
        "tags": ["Shopping Cart"],
        "parameters": [
          {
            "name": "cartItemId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CartItemQuantityRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Cart item quantity updated successfully"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "summary": "Delete a cart item by ID",
        "tags": ["Shopping Cart"],
        "parameters": [
          {
            "name": "cartItemId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cart item deleted successfully"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "CartItemRequestDto": {
        "type": "object",
        "properties": {
          "bookId": {
            "type": "integer"
          },
          "quantity": {
            "type": "integer",
            "minimum": 1,
            "maximum": 100
          }
        },
        "required": ["bookId", "quantity"]
      },
      "CartItemQuantityRequestDto": {
        "type": "object",
        "properties": {
          "quantity": {
            "type": "integer",
            "minimum": 1,
            "maximum": 100
          }
        },
        "required": ["quantity"]
      },
      "ShoppingCartDto": {
        "type": "object",
        "properties": {
        }
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}